Add files via upload
These are the final files used in the program. Main sales Graph and CovidGraphs are the finalized versions of the previously uploaded code. Where salesgraph contains the files faisal uploaded and CovidGraphs contains the files that Yossouf and Joe uploaded. While the uploading data and graphs via user selection along with the user interface was done by me.